RefSR-NeRF/requirements.txt
asttokens
astunparse
cycler
decorator
fonttools
imageio[ffmpeg]
joblib
kiwisolver
matplotlib
mindspore-dev==2.0.0.dev20230226
numpy
opencv-python
packaging
pandas
Pillow
protobuf
psutil
PyMCubes==0.1.2
pyparsing
python-dateutil
pytz
PyYAML
scikit-learn
scipy
six
sklearn
threadpoolctl
tqdm
trimesh==3.20.2
typing_extensions
